Fix: ../src/slic3r/GUI/GLModel.cpp:180:36: warning: comparison of integer expressions of different signedness: 'int' and 'std::vector<Slic3r::GUI::GLModel::RenderData>::size_type' {aka 'long unsigned int'} [-Wsign-compare]

+++ src/slic3r/GUI/GLModel.cpp
@@ -175,7 +175,7 @@ void GLModel::set_color(int entity_id, const std::array<float, 4>& color)
     }
 }
 
-const std::array<float, 4> GLModel::get_color(int entity_id) const
+const std::array<float, 4> GLModel::get_color(size_t entity_id) const
 {
     if (entity_id < 0 || entity_id >= m_render_data.size()) return {};
     return m_render_data[entity_id].color;
